# backend/internal/httpapi/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.History:
    properties:
      last_updated_at:
        type: string
      next_run:
        type: string
      samples:
        items:
          $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.PriceSample'
        type: array
    type: object
...
  github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.PriceSample:
    properties:
      date:
        description: YYYY-MM-DD
        type: string
      max_minor:
        type: integer
      median_minor:
        type: integer
      min_minor:
        type: integer
      n_quotes:
        type: integer
    type: object
  github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.PriceSummary:
    properties:
      currency:
        type: string
      delta_pct:
        type: number
      percentile:
        type: number
      price_minor:
        type: integer
    type: object
...
  github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Trip:
    properties:
      created_at:
        type: string
      cron_expr:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      legs:
        description: |-
          Legs is populated by the list/get handler from a separate
          ListWatches call — Trip itself carries no leg data.
        items:
          $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Watch'
        type: array
      name:
        type: string
      timezone:
        type: string
    type: object
  github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Watch:
    properties:
      consecutive_failures:
        type: integer
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      kind:
        type: string
      last_error:
        type: string
      last_updated_at:
        type: string
      name:
        type: string
      next_run:
        type: string
      params:
        items:
          type: integer
        type: array
      price:
        $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.PriceSummary'
      stale:
        type: boolean
      threshold_pct:
        type: number
      trip_id:
        type: string
    type: object
...
  internal_httpapi_handlers.runNowResponse:
    properties:
      message:
        type: string
    type: object
  internal_httpapi_handlers.tripRequest:
    properties:
      cron_expr:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      timezone:
        type: string
    type: object
  internal_httpapi_handlers.watchRequest:
    properties:
      enabled:
        type: boolean
      kind:
        type: string
      name:
        type: string
      params:
        items:
          type: integer
        type: array
      threshold_pct:
        type: number
      trip_id:
        type: string
    type: object
info:
  contact: {}
  description: |-
    Backend for the Pricewatch admin panel — watches, price history, run logs, settings, and Discord channel status. Single-admin: every route except /auth/login and /meta requires a valid session.
    Versioned by URL prefix (/api/v1/...) — see internal/httpapi/handlers.APIVersion and internal/httpapi/router.go.
  title: Pricewatch API
  version: "1.0"
paths:
...
  /trips:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Trip'
            type: array
      security:
      - CookieAuth: []
      summary: List trips, each with its legs
      tags:
      - trips
    post:
      consumes:
      - application/json
      parameters:
      - description: Trip fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi_handlers.tripRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Trip'
        "400":
          description: invalid request body, or a validation error naming the bad
            field
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Create a trip
      tags:
      - trips
  /trips/{id}:
    delete:
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: no content
        "404":
          description: trip not found
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Delete a trip and every leg in it
      tags:
      - trips
    get:
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Trip'
        "404":
          description: trip not found
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Get a trip and its legs
      tags:
      - trips
    patch:
      consumes:
      - application/json
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      - description: Full trip fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi_handlers.tripRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Trip'
        "404":
          description: trip not found
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Replace a trip
      tags:
      - trips
  /watches:
    get:
      description: Every watch, each enriched with its current price, freshness, and
        staleness state.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Watch'
            type: array
        "500":
          description: list watches
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: List watches
      tags:
      - watches
    post:
      consumes:
      - application/json
      description: Enabled by default. Triggers a background backfill fetch immediately
        so the chart isn't empty for a week, and reloads the scheduler so it's picked
        up on the next cron cycle without a restart.
      parameters:
      - description: Watch fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi_handlers.watchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Watch'
        "400":
          description: invalid request body, or a validation error naming the bad
            field
          schema:
            type: string
        "500":
          description: create watch
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Create a watch
      tags:
      - watches
  /watches/{id}:
    delete:
      parameters:
      - description: Watch ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: no content
        "404":
          description: watch not found
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Delete a watch
      tags:
      - watches
    get:
      parameters:
      - description: Watch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Watch'
        "404":
          description: watch not found
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Get a watch
      tags:
      - watches
    patch:
      consumes:
      - application/json
      description: Full-replace semantics despite the PATCH verb — the panel's watch-builder
        form always submits every field, so there's no partial-merge case to support.
      parameters:
      - description: Watch ID
        in: path
        name: id
        required: true
        type: string
      - description: Full watch fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi_handlers.watchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.Watch'
        "400":
          description: invalid request body, or a validation error naming the bad
            field
          schema:
            type: string
        "404":
          description: watch not found
          schema:
            type: string
        "500":
          description: update watch
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Replace a watch
      tags:
      - watches
  /watches/{id}/history:
    get:
      description: Server-side cached for exactly as long as the watch's own schedule
        (see internal/httpapi.historyCache) — "data only changes when the cron fires."
      parameters:
      - description: Watch ID
        in: path
        name: id
        required: true
        type: string
      - default: 90d
        description: Lookback window, e.g. 30d/90d/365d
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SamPariat_pricewatch_internal_httpapi_presenter_v1.History'
        "500":
          description: list price samples
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Get a watch's price history
      tags:
      - watches
  /watches/{id}/run:
    post:
      description: Runs the pipeline immediately, outside the normal schedule, bypassing
        the provider cache so it always fetches fresh (see providers.SkipCache). Honours
        the global dry-run setting for whether the result actually gets sent to Discord,
        matching the scheduled path's own behavior.
      parameters:
      - description: Watch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi_handlers.runNowResponse'
        "404":
          description: watch not found
          schema:
            type: string
        "502":
          description: 'run failed: <cause>'
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Run a watch now
      tags:
      - watches
securityDefinitions:
  CookieAuth:
    description: 'httpOnly session cookie issued by POST /auth/login (see internal/httpapi/middleware.Session).
      Paste as "Cookie: pw_session=<value>" — this is the real wire format, not an
      approximation of the cookie mechanism.'
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"